+++ server/src/test/java/com/continuuity/loom/store/node/SQLNodeStoreTest.java
@@ -21,6 +21,7 @@ import com.continuuity.loom.common.conf.Constants;
 import com.continuuity.loom.common.conf.guice.ConfigurationModule;
 import com.continuuity.loom.store.DBHelper;
 import com.continuuity.loom.store.guice.StoreModule;
+import com.continuuity.loom.store.guice.TestStoreModule;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
 import org.junit.AfterClass;
@@ -42,7 +43,7 @@ public class SQLNodeStoreTest extends NodeStoreTest {
     sqlConf.setLong(Constants.ID_START_NUM, 1);
     sqlConf.setLong(Constants.ID_INCREMENT_BY, 1);
     Injector injector = Guice.createInjector(new ConfigurationModule(sqlConf),
-                                             new StoreModule(),
+                                             new TestStoreModule(),
                                              new CodecModules().getModule());
     sqlNodeStoreService = injector.getInstance(SQLNodeStoreService.class);
     sqlNodeStoreService.startAndWait();